# Fix issue where the last letter of self-send messages would get dropped